Port horwitz keymap from firmware23 to firmware25 (build + docs)
Keyboard path: keyboards/moonlander/ → keyboards/zsa/moonlander/reva
Build target:  moonlander:horwitz → zsa/moonlander/reva:horwitz
Tests (make test:horwitz): 29 passing, unchanged.

Fixes to QMK/ZSA infrastructure (non-keymap files):
- lib/python/qmk/math.py: ast.Num → ast.Constant (removed in Python 3.12)
- builddefs/build_test.mk: add LDFLAGS/CFLAGS for Apple clang 21 test
  builds (-Wno-include-next-absolute-path, -lstdc++, -lpthread);
  add $(TEST)_VPATH hook so test.mk can append include dirs after
  tests/test_common in VPATH
- platforms/chibios/platform.mk: bridge OSLIBINC/LIBINC name mismatch
  (oslib.mk sets LIBINC; platform.mk still references OSLIBINC)
- quantum/process_keycode/process_space_cadet.c: () → (void)
- tmk_core/protocol/usb_device_state.c: () → (void)
  (both: -Wstrict-prototypes, Apple clang 21)
- tests/test_common/keyboard_report_util.hpp: static_cast<uint8_t>
  narrowing fix

Keymap changes (keyboards/zsa/moonlander/keymaps/horwitz/):
- keymap.json: new file declaring zsa/defaults community module (owns
  LED_LEVEL and TOGGLE_LAYER_COLOR in firmware25)
- rules.mk: remove ORYX_ENABLE and RGB_MATRIX_CUSTOM_KB (orphaned from
  firmware23; oryx is now a community module and horwitz uses no oryx
  APIs; RGB_MATRIX_CUSTOM_KB dropped the mcp23018 linker dependency);
  SRC = → SRC += so keyboard-level mcp23018.c is not overwritten
- keymap.c: ML_SAFE_RANGE → ZSA_SAFE_RANGE; guard rawhid_state
  references with #ifdef COMMUNITY_MODULE_ORYX_ENABLE; key_overrides
  from NULL-terminated pointer to array (firmware25 uses ARRAY_SIZE)
- run-cppcheck.sh, run-clang-tidy.sh: update paths to zsa/moonlander
- DEVELOPMENT.md: update all paths/commands; add firmware23→25 porting
  notes documenting every non-obvious migration decision

.gitignore: add negation rules for keyboards/zsa/**/keymaps/horwitz

+++ tests/test_common/keyboard_report_util.hpp
@@ -35,5 +35,5 @@ private:
 
 template<typename... Ts>
 inline testing::Matcher<report_keyboard_t&> KeyboardReport(Ts... keys) {
-    return testing::MakeMatcher(new KeyboardReportMatcher(std::vector<uint8_t>({keys...})));
+    return testing::MakeMatcher(new KeyboardReportMatcher(std::vector<uint8_t>({static_cast<uint8_t>(keys)...})));
 }
